M7 RAG_Retrieval Augmented Generation/Context Management & Langchain & API/Langchain_Chat with Memory/Langchain_ Chat/requirements.txt
langchain>=0.2
langchain-openai>=0.1
python-dotenv>=1.0
openai>=1.0
fastapi>=0.110
uvicorn>=0.27
gradio>=4.0
requests>=2.31